# RuGram/docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.CreatePostRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      image_url:
        type: string
      status:
        enum:
        - active
        - draft
        - archived
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
      user_id:
        type: string
    required:
    - title
    - user_id
    type: object
...
  dto.MetaData:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
  dto.PaginationResponse:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/dto.MetaData'
    type: object
  dto.PostResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      likes_count:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  dto.UpdatePostRequest:
    properties:
      description:
        maxLength: 1000
        type: string
      image_url:
        type: string
      status:
        enum:
        - active
        - draft
        - archived
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    type: object
...
  utils.ErrorResponseStruct:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  utils.StandardResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:4200
info:
  contact:
    email: support@rugram.com
    name: Gleb, Pavel
  description: API for RuGram social platform with MongoDB, JWT, OAuth, Redis.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: RuGram API
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of posts
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        maximum: 100
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginationResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: List posts with pagination
      tags:
      - Posts
    post:
      consumes:
      - application/json
      description: Creates a post with given data
      parameters:
      - description: Post data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PostResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Create a new post
      tags:
      - Posts
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-deletes a post by ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Delete a post (soft delete)
      tags:
      - Posts
    get:
      consumes:
      - application/json
      description: Returns a single post by its ID
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PostResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Get post by ID
      tags:
      - Posts
    patch:
      consumes:
      - application/json
      description: Updates post fields (full or partial)
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PostResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - Posts
    put:
      consumes:
      - application/json
      description: Updates post fields (full or partial)
      parameters:
      - description: Post ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PostResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Update a post
      tags:
      - Posts
  /posts/user/{userId}:
    get:
      consumes:
      - application/json
      description: Returns paginated posts of a specific user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        maximum: 100
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.PaginationResponse'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.ErrorResponseStruct'
      security:
      - BearerAuth: []
      summary: Get posts by user ID
      tags:
      - Posts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"